requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
psycopg2-binary

# Auth dependencies
passlib[bcrypt]
python-jose[cryptography]

# Vision Model Dependencies
torch
torchvision
transformers
accelerate
pillow
python-multipart
